start more kernel stuff

+++ src/kernel/code_2FE40.h
@@ -3,7 +3,7 @@
 
 #include <ultra64.h>
 
-void uvComputeLineEndP(s16*, s16*, u16*, f32, f32, f32, f32, f32, f32, f32, s32, f32, f32);
-s32 func_8022F62C(s32, s32, s32, s32);
+void uvComputeLineEndP(s16*, s16*, u16*, f32, f32, f32, f32, f32, f32, f32, u8, f32, f32);
+s32 func_8022F62C(s32, s16, s16, u8);
 
 #endif // CODE_2FE40_H
